chore: merge files in cloudfront
